+++ src/Logger.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Tivins\LlmBasic;
-
-class Logger
-{
-
-}